rename Header to PageTitle
