+++ src/app/ui/Page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Header from './Header';
+import PageTitle from './PageTitle';
 import Nav from './nav/Nav';
 
 import styles from './Page.module.scss';
@@ -9,7 +9,7 @@ export default function Page(Component: React.ComponentType)  {
   const component = () => {
     return (
       <div className={styles.Page}>
-        <Header />
+        <PageTitle />
         <Nav/>
         <Component />
       </div>
