+++ src/app/ui/PageTitle.module.scss
@@ -1,4 +1,4 @@
-.Header {
+.PageTitle {
   width: 100px;
   padding: 30px;
   position: absolute;
